fix: code review - moved mail/util.js to libraries/mail-util.js
유틸파일 라이브러리로 이동
